+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/CorsCodeGenerator.kt
@@ -55,7 +55,7 @@ class CorsCodeGenerator : AbstractKotlinCodeGenerator() {
             )
             resp.addHeader(
                 "Access-Control-Allow-Headers",
-                "Content-Type, Authorization, Content-Length,X-Requested-With,Authorization"
+                "Content-Type, Authorization, Content-Length,X-Requested-With"
             )
             chain.doFilter(req, resp)
         """.trimIndent()