+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/CorsCodeGeneratorTest.kt
@@ -58,7 +58,7 @@ internal class CorsCodeGeneratorTest {
                     )
                     resp.addHeader(
                         "Access-Control-Allow-Headers",
-                        "Content-Type, Authorization, Content-Length,X-Requested-With,Authorization"
+                        "Content-Type, Authorization, Content-Length,X-Requested-With"
                     )
                     chain.doFilter(req, resp)
                   }