+++ src/test/kotlin/com/wutsi/codegen/kotlin/server/ServerCodeGeneratorTest.kt
@@ -7,6 +7,7 @@ import com.wutsi.codegen.core.generator.CodeGenerator
 import com.wutsi.codegen.editorconfig.EditorConfigCodeGenerator
 import com.wutsi.codegen.github.GitCodeGenerator
 import com.wutsi.codegen.kotlin.KotlinMapper
+import com.wutsi.codegen.renovate.RenovateCodeGenerator
 import io.swagger.v3.oas.models.OpenAPI
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
@@ -37,7 +38,7 @@ internal class ServerCodeGeneratorTest {
     fun `generators`() {
         val codegen = ServerCodeGenerator(mapper = KotlinMapper(context))
 
-        assertEquals(18, codegen.generators.size)
+        assertEquals(16, codegen.generators.size)
         assertTrue(codegen.generators[0] is ServerDtoCodeGenerator)
         assertTrue(codegen.generators[1] is ServerDelegateCodeGenerator)
         assertTrue(codegen.generators[2] is ServerControllerCodeGenerator)
@@ -48,13 +49,11 @@ internal class ServerCodeGeneratorTest {
         assertTrue(codegen.generators[7] is EditorConfigCodeGenerator)
         assertTrue(codegen.generators[8] is ServerGithubActionsCodeGenerator)
         assertTrue(codegen.generators[9] is GitCodeGenerator)
-        assertTrue(codegen.generators[10] is CacheCodeGenerator)
-        assertTrue(codegen.generators[11] is MQueueCodeGenerator)
-        assertTrue(codegen.generators[12] is SwaggerCodeGenerator)
-        assertTrue(codegen.generators[13] is ServerReadmeCodeGenerator)
-        assertTrue(codegen.generators[14] is CorsCodeGenerator)
-        assertTrue(codegen.generators[15] is TracingCodeGenerator)
-        assertTrue(codegen.generators[16] is DatabaseCodeGenerator)
-        assertTrue(codegen.generators[17] is SecurityCodeGenerator)
+        assertTrue(codegen.generators[10] is SwaggerCodeGenerator)
+        assertTrue(codegen.generators[11] is ServerReadmeCodeGenerator)
+        assertTrue(codegen.generators[12] is CorsCodeGenerator)
+        assertTrue(codegen.generators[13] is DatabaseCodeGenerator)
+        assertTrue(codegen.generators[14] is SecurityCodeGenerator)
+        assertTrue(codegen.generators[15] is RenovateCodeGenerator)
     }
 }