+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # See https://keepachangelog.com/en/1.0.0/
 
+## [0.0.32] 2021-05-07
+### Added
+- Auto increase version of SDK
+- Add support for renovate
+- Add `Content-Type` header on SDK endpoints
+- Integration of `wutsi-platform` and simplification of dependecies
+### Changed
+- Update dependency version
+
 ## [0.0.31] 2021-05-02
 ### Added
 - Mapping of list of primitives